fix: harden checkpoint parent link invariants
